Revert "update geneticAlgorithmKernel function"
This reverts commit dfd953fe9ee5c7a5d1802864d88577671ad5c6c7.

+++ Selections/RouleteWheel.h
@@ -5,10 +5,11 @@
 #include <curand_kernel.h>
 
 namespace tsp {
+
 	__device__
-		int rouletteWheelSelection(int* fitness, int populationSize, curandState* state, float totalFitness) {
-		int slice = (int)(curand_uniform(state) * totalFitness);
-		int total = 0;
+	int rouletteWheelSelection(float* fitness, int populationSize, curandState* state, float totalFitness) {
+		float slice = curand_uniform(state) * totalFitness;
+		float total = 0;
 		for (int i = 0; i < populationSize; ++i) {
 			total += fitness[i];
 			if (total > slice) {
@@ -17,6 +18,7 @@ namespace tsp {
 		}
 		return populationSize - 1;
 	}
+
 }
 
-#endif
+#endif
